Built, not yet tested

+++ guest.sh
@@ -4,26 +4,32 @@ export HOME=/root
 export PACKAGES="/tmp/packages"
 
 
+PURGE_UNNEEDED=false
 ## Purge unnecessary packages
-export PURGE=`cat $PACKAGES/purge.list`
-echo "Purging unneeded packages"
-sudo apt-get purge $PURGE --assume-yes 
-
-## Install new packages
-if $UPDATE_APT ; then
-  echo "Adding universe apt repo..."
-  sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu $(lsb_release -sc) universe multiverse" 
-  sudo apt-get update 
+if $PURGE_UNNEEDED ; then
+  export PURGE=`cat $PACKAGES/purge.list`
+  echo "Purging unneeded packages"
+  sudo apt-get purge $PURGE --assume-yes 
 fi
-cd $PACKAGES
-for package in `ls -d */ | sed 's/\///'` ; do
-  echo "Installing $package"
-  $package/install.sh
-  if [ $? -ne 0 ] ;  then
-    echo "$package failed to install"
-    exit 1
+
+ADD_NEW=false
+## Install new packages
+if $ADD_NEW ; then
+  if $UPDATE_APT ; then
+    echo "Adding universe apt repo..."
+    sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu $(lsb_release -sc) universe multiverse" 
+    sudo apt-get update 
   fi
-done
+  cd $PACKAGES
+  for package in `ls -d */ | sed 's/\///'` ; do
+    echo "Installing $package"
+    $package/install.sh
+    if [ $? -ne 0 ] ;  then
+      echo "$package failed to install"
+      exit 1
+    fi
+  done
+fi
 
 ## Clean up
 echo "Cleaning up.."
